+++ composeApp/src/commonMain/kotlin/plus/vplan/app/data/repository/TeacherRepositoryImpl.kt
@@ -1,5 +1,6 @@
 package plus.vplan.app.data.repository
 
+import co.touchlab.kermit.Logger
 import io.ktor.client.HttpClient
 import io.ktor.client.request.get
 import io.ktor.client.statement.bodyAsText
@@ -24,6 +25,7 @@ import plus.vplan.app.data.source.network.toErrorResponse
 import plus.vplan.app.domain.cache.CacheState
 import plus.vplan.app.domain.data.Response
 import plus.vplan.app.domain.model.School
+import plus.vplan.app.domain.model.SchoolApiAccess
 import plus.vplan.app.domain.model.Teacher
 import plus.vplan.app.domain.repository.TeacherRepository
 import plus.vplan.app.utils.sendAll
@@ -71,24 +73,38 @@ class TeacherRepositoryImpl(
             }
             send(CacheState.Loading(id.toString()))
 
-            val accessResponse = httpClient.get("${api.url}/api/v2.2/teacher/$id")
-            if (accessResponse.status == HttpStatusCode.NotFound && accessResponse.isResponseFromBackend()) {
-                vppDatabase.teacherDao.deleteById(listOf(id))
-                return@channelFlow send(CacheState.NotExisting(id.toString()))
+            val existing = vppDatabase.teacherDao.getById(id).first()
+            var schoolApiAccess: SchoolApiAccess? = null
+            if (existing != null) {
+                schoolApiAccess = existing.schoolId.let { vppDatabase.schoolDao.findById(it).first()?.toModel()?.getSchoolApiAccess() }
             }
-            if (!accessResponse.status.isSuccess()) return@channelFlow send(CacheState.Error(id.toString(), accessResponse.toErrorResponse<Teacher>()))
-            val accessData = ResponseDataWrapper.fromJson<TeacherUnauthenticatedResponse>(accessResponse.bodyAsText())
-                ?: return@channelFlow send(CacheState.Error(id.toString(), Response.Error.ParsingError(accessResponse.bodyAsText())))
-
-            val school = vppDatabase.schoolDao.findById(accessData.schoolId).first()?.toModel()
-                .let {
-                    if (it is School.IndiwareSchool && !it.credentialsValid) return@channelFlow send(CacheState.Error(id.toString(), Response.Error.Other("no school for teacher $id")))
-                    if (it?.getSchoolApiAccess() == null) return@channelFlow send(CacheState.Error(id.toString(), Response.Error.Other("no school for teacher $id")))
-                    it.getSchoolApiAccess()!!
+            if (schoolApiAccess == null) {
+                val accessResponse = httpClient.get("${api.url}/api/v2.2/teacher/$id")
+                if (accessResponse.status == HttpStatusCode.NotFound && accessResponse.isResponseFromBackend()) {
+                    vppDatabase.teacherDao.deleteById(listOf(id))
+                    return@channelFlow send(CacheState.NotExisting(id.toString()))
                 }
+                if (!accessResponse.status.isSuccess()) return@channelFlow send(CacheState.Error(id.toString(), accessResponse.toErrorResponse<Teacher>()))
+                val accessData = ResponseDataWrapper.fromJson<TeacherUnauthenticatedResponse>(accessResponse.bodyAsText())
+                    ?: return@channelFlow send(CacheState.Error(id.toString(), Response.Error.ParsingError(accessResponse.bodyAsText())))
+
+                schoolApiAccess = vppDatabase.schoolDao.findById(accessData.schoolId).first()?.toModel()
+                    .let {
+                        if (it is School.IndiwareSchool && !it.credentialsValid) return@channelFlow send(CacheState.Error(id.toString(), Response.Error.Other("no school for teacher $id")))
+                        if (it?.getSchoolApiAccess() == null) return@channelFlow send(CacheState.Error(id.toString(), Response.Error.Other("no school for teacher $id")))
+                        it.getSchoolApiAccess()
+                    }
+            }
+
+            if (schoolApiAccess == null) {
+                Logger.i { "No school to update teacher $id" }
+                vppDatabase.teacherDao.deleteById(listOf(id))
+                trySend(CacheState.NotExisting(id.toString()))
+                return@channelFlow
+            }
 
             val response = httpClient.get("${api.url}/api/v2.2/teacher/$id") {
-                school.authentication(this)
+                schoolApiAccess.authentication(this)
             }
             if (!response.status.isSuccess()) return@channelFlow send(CacheState.Error(id.toString(), response.toErrorResponse<Teacher>()))
             val data = ResponseDataWrapper.fromJson<TeacherItemResponse>(response.bodyAsText())
